Batch A: Remove redundant usings, casts, qualifications, assignments
Apply auto-fixable informational style diagnostics:
- IDE0005: remove unnecessary usings (45 files)
- IDE0004: remove unnecessary casts (+20 files)
- IDE0001: simplify name (+22 files)
- IDE0002: simplify member access (+10 files)
- IDE0003: remove this. qualification (+0 files — already applied by IDE0002)
- IDE0059: remove unnecessary value assignments (+13 files)

110 files total, +351/-417 lines. Strictly subtractive.
0 warnings, 0 errors after build.

+++ Tests/Opc.Ua.Client.Tests/FileSystem/FileSystemClientMetadataTests.cs
@@ -52,7 +52,7 @@ namespace Opc.Ua.Client.Tests.FileSystem
                 Size = 12_345UL,
                 Writable = true,
                 UserWritable = false,
-                OpenCount = (ushort)3
+                OpenCount = 3
             };
             props.Realize();
             harness.RegisterFile(harness.Root, new QualifiedName("data.bin"), properties: props);
@@ -80,7 +80,7 @@ namespace Opc.Ua.Client.Tests.FileSystem
                 Size = 0UL,
                 Writable = true,
                 UserWritable = true,
-                OpenCount = (ushort)0,
+                OpenCount = 0,
                 MimeType = "application/json",
                 MaxByteStringLength = 64_000u,
                 LastModifiedTime = expectedModified
